[MSP430] WIP binary output.
This adds the base infrastructure and so on to output from MSP430 as binary.

The implementation is not satisfactory at present, since the TD files won't generate the disassembler etc.
This appears to be because some of the bit patterns implied by the core classes are not orthogonal, it should be possible to re-work what's here to achieve that.