[PPC,Darwin] Darwin's default relocation model is PIC.
I.E. not mdynamic-no-pic.

Unfortunately, there is significant test-suite fallout from this.  Some fixable by forcing the relocation model back to dynamic-no-pic.  Some is not easily fixed this way and the intent of the test-cases is not 100% clear, so these have been XFAILed for now.

These are old-style llc | grep and the target lines are not specific, so it's hard to know what to do.

* mem_update.ll
* tailcall1-64.ll
* tailcall1.ll

In any case, the tail calls could well be broken by PIC code.