[llvm, ppc] Revise “Darwin Directive” to “MCPU”.
‘Darwin Directive’ really has very little to do with OS X / Darwin specifically.  It is mostly about having a mechanism to test for functionality that depends on the MCPU value without repeated string tests.

The one ‘darwin specific’ aspect is that the .machine directives in older darwin assembler files are like “ppc7400” where the current style is like “7400”.  We cater for parsing legacy assembler files, but switch to using the new form for written files (the current assembler output from LLVM is already incompatible with cctools ‘as’).

The MCPU values and strings have been re-synced.  Note that ELF platforms do not currently act on the MCPU value in .machine directives (except to parse and ignore “any” “push” and “pop”).